requirements.txt
webuiapi
pillow
dghs-imgutils>=0.4.6
scikit-image
pandas
onnxruntime-gpu
httpx>=0.23.0
gradio==4.21.0
hbutils
click>=8
di-toolkit>=0.2
pyarrow